ath11k: move ring mask variables to core.c
It should have been in core.c in the first place as the declarations are in
core.h. And these are needed by both pci.c and ahb.c so there needs to be a
common place for them.

Also add ath11k_ prefix to rx_mon_status_ring_mask to make it easy add
EXPORT_SYMBOL() later.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Patchwork-Id: 11541355